requirements.txt
django==1.11.5
django-mptt==0.9.0
pinax-referrals==3.0.3
django-phonenumber-field==2.0.0
django-user-accounts==2.0.3
celery==4.1.1
redis==2.10.6
psycopg2==2.7.5
django-bootstrap-datepicker-plus==3.0.4
django-widget-tweaks==1.4.2